Added data transfer object
