+++ lib/models/task_dto.dart
@@ -1,0 +1,9 @@
+class TaskDTO {
+  String name;
+  String description;
+  int durationWork;
+  int durationBreak;
+  int totalTime = 0;
+  int goalTime;
+  String category;
+}
